Delaying the writes for toosmallthreshold tests
The writes are so fast that files get the same date-time stamp so that it's not possible to validate the files in order to check that no logs got lost.

Writing less files at a much slower speed to ensure that even with the rest of the tests in play we always get distinct modified times that we can order the files by.